% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_smbo.R
\name{algorithm_SMBO}
\alias{algorithm_SMBO}
\title{Title}
\usage{
algorithm_SMBO(modelOptimizer, distanceFunction, designGenerator,
  budget = 100, useBP = F, useEI = F)
}
\arguments{
\item{nDim}{}
}
\description{
Title
}
