% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_SpotOptimByName.R
\name{algorithm_SpotOptimByName}
\alias{algorithm_SpotOptimByName}
\title{Title}
\usage{
algorithm_SpotOptimByName(strAlgoName, lower, upper, control, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
