% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_cego.R
\name{cego_CreationFunctionBinary}
\alias{cego_CreationFunctionBinary}
\title{Title}
\usage{
cego_CreationFunctionBinary(nDim)
}
\arguments{
\item{nDim}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
