% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_cego.R
\name{cego_CreationFunctionReal}
\alias{cego_CreationFunctionReal}
\title{Title}
\usage{
cego_CreationFunctionReal(nDim, lBounds = rep(-5, nDim), uBounds = rep(5,
  nDim))
}
\arguments{
\item{nDim}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
