% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyTimePlot.R
\name{fast20ResDF}
\alias{fast20ResDF}
\title{Compile the 'batchtools' result list into a data.frame}
\usage{
fast20ResDF(resultList)
}
\arguments{
\item{resultList}{the list you get by ideaLoadResultList}
}
\value{
resultDF
}
\description{
Currently uses FR format for FR's plots, ask me ;-)
}
