% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{ideaLoadResultList}
\alias{ideaLoadResultList}
\title{Load experiment results into result list}
\usage{
ideaLoadResultList(reg, doDelete = T, waitJobs = T)
}
\arguments{
\item{reg}{The registry from which the results shall be loaded}
}
\description{
Load experiment results into result list
}
