% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{ideaPathIsBaseDir}
\alias{ideaPathIsBaseDir}
\title{ideaPathIsBaseDir}
\usage{
ideaPathIsBaseDir(path, sess = NULL)
}
\arguments{
\item{path}{path to the folder}

\item{sess}{SSH session}
}
\description{
Check if a path is really a registry dear before you delete it etc.
}
