% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{ideaTutorial}
\alias{ideaTutorial}
\title{Run or list ideaTutorials}
\usage{
ideaTutorial(index = NULL)
}
\arguments{
\item{index}{Integer, index of which tutorial to run. If NULL (default) all tutorials will be listed.}
}
\description{
There are some tutorials for the usage of the ideaBatch and batchtools package available.
You can list them by simply typing 'ideaTutorial()'. You will get a list of available tutorials.
By specifying the tutorial index you can open a document: e.g. 'ideaTutorial(1)'
}
