% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{initBatchTools}
\alias{initBatchTools}
\title{Initiate ideaBatch package}
\usage{
initBatchTools()
}
\description{
Initiate the ideaBatch package, install all requirements on the cluster copy files,
create your user config.
}
