% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{insertPathToTemplate}
\alias{insertPathToTemplate}
\title{Insert Variable into file}
\usage{
insertPathToTemplate(fullPath, paramVector)
}
\description{
Scans thorugh a file looking for codes in the form of e.g. #1# or #2# etc.
Parameters are filled from paramVector into the file
}
\keyword{internal}
