% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterFunctionsIDEASlurm.R
\name{makeClusterFunctionsIDEASlurm}
\alias{makeClusterFunctionsIDEASlurm}
\title{Title}
\usage{
makeClusterFunctionsIDEASlurm(template = "slurm", array.jobs = TRUE,
  nodename = "localhost", scheduler.latency = 1, fs.latency = 65)
}
\description{
Title
}
