% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{prepareBaseFiles}
\alias{prepareBaseFiles}
\title{prepareBaseFiles}
\usage{
prepareBaseFiles()
}
\description{
Copies all required files from the package installation folder to the user specified foder
}
