% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalRanking.R
\name{rankDataByPostHocKruskalConover}
\alias{rankDataByPostHocKruskalConover}
\title{rankDataByPostHocKruskalConover}
\usage{
rankDataByPostHocKruskalConover(results, methods)
}
\arguments{
\item{results}{vector of endresults}

\item{methods}{vector of strings, method that was used to create the results}
}
\value{
list with ranks
}
\description{
Use posthoc.kruskal.conover.test to statistically rank the given methods.
}
