% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{readSSHInfo}
\alias{readSSHInfo}
\title{Parse ssh-config file}
\usage{
readSSHInfo(nodeName)
}
\arguments{
\item{nodeName}{string with the name to search for}
}
\value{
server info list
}
\description{
Scans through the .ssh/config file looking for a configuration with the given nodeName
}
\keyword{internal}
