% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{synchronizeFolder}
\alias{synchronizeFolder}
\title{Synchronize Experiment Folder}
\usage{
synchronizeFolder(doDelete = T)
}
\arguments{
\item{dir}{the dir that is synchronized}
}
\description{
Synchronizes the Experiment folder to the cluster
}
