% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{synchronizeJobsFolder}
\alias{synchronizeJobsFolder}
\title{Synchronize Experiment Folder}
\usage{
synchronizeJobsFolder(jc)
}
\arguments{
\item{dir}{the dir that is synchronized}
}
\description{
Synchronizes the Experiment folder to the cluster
}
