% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyTimePlot.R
\name{anyTimePerformancePlot}
\alias{anyTimePerformancePlot}
\title{Title}
\usage{
anyTimePerformancePlot(data, confidenceInterval = 0.95, yLim = NULL,
  xLim = NULL, ylog = F, xlog = F, useMinMax = F)
}
\description{
Title
}
